Moved user to a different module
