+++ src/lib.rs
@@ -1,21 +1,19 @@
 #![feature(iterator_try_collect)]
 #![warn(
-anonymous_parameters,
-missing_copy_implementations,
-missing_debug_implementations,
-rust_2018_idioms,
-rustdoc::private_doc_tests,
-trivial_casts,
-trivial_numeric_casts,
-unused,
-future_incompatible,
-nonstandard_style,
-unsafe_code,
-unused_import_braces,
-unused_results,
-variant_size_differences
+    anonymous_parameters,
+    missing_copy_implementations,
+    missing_debug_implementations,
+    rust_2018_idioms,
+    rustdoc::private_doc_tests,
+    trivial_casts,
+    trivial_numeric_casts,
+    unused,
+    future_incompatible,
+    nonstandard_style,
+    unsafe_code,
+    unused_import_braces,
+    unused_results,
+    variant_size_differences
 )]
 
-
 pub mod service;
-
