+++ src/main.rs
@@ -1,6 +1,6 @@
+use dotenv::dotenv;
 use std::env;
 use tokio::io;
-use dotenv::dotenv;
 
 #[tokio::main]
 async fn main() -> Result<(), anyhow::Error> {
